4847 Move integration tests to client AWS package
